+++ database/models/product.js
@@ -7,7 +7,7 @@ const Product = (sequelize, DataTypes) => {
       type: DataTypes.INTEGER
     },
     name: DataTypes.STRING,
-    price: DataTypes.DOUBLE,
+    price: DataTypes.DECIMAL(10, 2),
     quantity: DataTypes.INTEGER
   }, {
     tableName: 'Products',
